Moved http socket server to index.js
