+++ index.js
@@ -2,6 +2,8 @@ const express = require('express');
 const cors = require('cors');
 const path = require('path');
 const fs = require('fs');
+const http = require('http');
+const socketIo = require('socket.io');
 const { google } = require('googleapis');
 const { oauth2Client, getAuthUrl, getTokensFromCode, setCredentialsFromEnv, listFilesInFolder } = require('./oauthHandler');
 const settingsRoutes = require('./routes/settingsRoute');
@@ -146,4 +148,57 @@ app.get('/oauth2callback', async (req, res) => {
   }
 });
 
-module.exports = app;
+// Create HTTP server with Socket.IO
+const server = http.createServer(app);
+const io = socketIo(server, {
+  cors: {
+    origin: config.server.corsOrigin,
+    methods: ["GET", "POST"]
+  }
+});
+
+// Socket.IO connection handling
+io.on('connection', (socket) => {
+  console.log('🔌 Client connected:', socket.id);
+  
+  socket.on('join-customer-room', () => {
+    socket.join(config.websocket.rooms.customer);
+    socket.join(`${config.defaultUsers.customer.email}-updates`);
+    console.log('👥 Customer joined update room');
+  });
+  
+  socket.on('join-admin-room', () => {
+    socket.join(config.websocket.rooms.admin);
+    socket.join(`${config.defaultUsers.admin.email}-updates`);
+    console.log('👤 Admin joined update room');
+  });
+  
+  socket.on('join-user-room', (userId) => {
+    socket.join(`${userId}-updates`);
+    console.log(`👤 User ${userId} joined their specific room`);
+  });
+  
+  socket.on('disconnect', () => {
+    console.log('🔌 Client disconnected:', socket.id);
+  });
+});
+
+// Make io available to routes
+app.set('io', io);
+
+// Test endpoint for WebSocket events
+app.post('/test-websocket', (req, res) => {
+  const io = req.app.get('io');
+  if (io) {
+    io.emit('test-event', {
+      message: 'Test WebSocket event',
+      timestamp: new Date().toISOString()
+    });
+    console.log('📡 Test WebSocket event emitted');
+    res.json({ message: 'Test event sent' });
+  } else {
+    res.status(500).json({ error: 'Socket.IO not available' });
+  }
+});
+
+module.exports = { app, server };
