+++ package.json
@@ -36,6 +36,7 @@
         "jest": "^30.0.5",
         "mongodb-memory-server": "^10.2.0",
         "nodemon": "^3.1.10",
+        "socket.io-client": "^4.7.4",
         "supertest": "^7.1.4"
     }
 }
